+++ src/controllers/CatchAllController.php
@@ -10,9 +10,10 @@ namespace venveo\redirect\controllers;
 
 use Craft;
 use craft\db\Paginator;
+use craft\helpers\AdminTable;
+use craft\helpers\Html;
 use craft\helpers\UrlHelper;
 use craft\web\Controller;
-use craft\web\Response;
 use venveo\redirect\Plugin;
 use venveo\redirect\records\CatchAllUrl;
 use yii\db\Query;
@@ -117,14 +118,11 @@ class CatchAllController extends Controller
         $this->requirePostRequest();
         $this->requireAcceptsJson();
 
-        $currentUser = Craft::$app->getUser()->getIdentity();
-        if (!$currentUser->can(Plugin::PERMISSION_MANAGE_404S)) {
-            return Craft::$app->response->setStatusCode('403', Craft::t('vredirect', 'You lack the required permissions to manage registered 404s'));
-        }
+        $this->requirePermission(Plugin::PERMISSION_MANAGE_404S);
 
         $data = \GuzzleHttp\json_decode(Craft::$app->request->getRawBody(), true);
         CatchAllUrl::deleteAll(['in', 'id', $data]);
-        return $this->asJson('Deleted');
+        return $this->asJson(['success' => true]);
     }
 
     public function actionIgnore()
@@ -156,5 +154,63 @@ class CatchAllController extends Controller
         $data = \GuzzleHttp\json_decode(Craft::$app->request->getRawBody(), true);
         CatchAllUrl::updateAll(['ignored' => false], ['in', 'id', $data]);
         return $this->asJson('Un-ignored');
+    }
+
+    public function actionHitsTable() {
+        $this->requireAcceptsJson();
+        $request = Craft::$app->getRequest();
+
+        $page = $request->getParam('page', 1);
+        $sort = $request->getParam('sort', null);
+        $limit = $request->getParam('per_page', 10);
+        $search = $request->getParam('search', null);
+        $offset = ($page - 1) * $limit;
+
+        $recordQuery = CatchAllUrl::find();
+
+        if ($search) {
+            $likeOperator = Craft::$app->getDb()->getIsPgsql() ? 'ILIKE' : 'LIKE';
+            $recordQuery->andWhere([
+                'or',
+                [$likeOperator, '[[id]]', $search],
+                [$likeOperator, '[[uri]]', $search],
+                [$likeOperator, '[[uid]]', $search],
+                [$likeOperator, '[[referrer]]', $search],
+                [$likeOperator, '[[dateUpdated]]', $search],
+                [$likeOperator, '[[dateCreated]]', $search]
+            ]);
+        }
+
+        if ($sort) {
+            $sortData = explode('|', $sort);
+            $sortKey = $sortData[0];
+            $sortDir = $sortData[1] === 'asc' ? SORT_ASC : SORT_DESC;
+            $orderParam = [$sortKey => $sortDir];
+            $recordQuery->orderBy($orderParam);
+        }
+
+        $total = $recordQuery->count();
+
+        $recordQuery->offset($offset);
+        $recordQuery->limit($limit);
+
+        $registered404s = $recordQuery->all();
+
+        $rows = [];
+        foreach ($registered404s as $customer) {
+            $rows[] = [
+                'id' => $customer['id'],
+                'title' => Html::encode($customer['uri']),
+                'referrer' => Html::encode($customer['referrer']),
+                'hitCount' => $customer['hitCount'],
+                'dateCreated' => $customer['dateCreated'],
+                'dateUpdated' => $customer['dateUpdated'],
+            ];
+        }
+
+        return $this->asJson([
+            'pagination' => AdminTable::paginationLinks($page, $total, $limit),
+            'data' => $rows,
+        ]);
     }
 }